CarTrip
This code calculates the amount of gasoline consumed per 100 km. It requests the distance in kilometers at the beginning and end, as well as the amount of gasoline at the start and end. It performs the mathematical calculation and provides the final result of the amount of gasoline consumed per 100 km.